move MLFF to __init__ to avoid circular import
